Document and wire up split view
