+++ features/apps/ui/src/main/kotlin/com/nasdroid/apps/ui/Navigation.kt
@@ -1,6 +1,7 @@
 package com.nasdroid.apps.ui
 
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.material3.windowsizeclass.WindowSizeClass
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -9,13 +10,14 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import androidx.navigation.navigation
 import com.nasdroid.apps.ui.discover.DiscoverAppsScreen
-import com.nasdroid.apps.ui.installed.InstalledAppsOverviewScreen
-import com.nasdroid.apps.ui.installed.logs.LogsScreen
+import com.nasdroid.apps.ui.installed.InstalledAppsScreen
+import com.nasdroid.apps.ui.installed.overview.logs.LogsScreen
 
 /**
  * Registers a nested navigation graph for the Apps feature.
  */
 fun NavGraphBuilder.appsGraph(
+    windowSizeClass: WindowSizeClass,
     navController: NavController,
     route: String,
     modifier: Modifier = Modifier,
@@ -26,7 +28,8 @@ fun NavGraphBuilder.appsGraph(
         route = route
     ) {
         composable("installed") {
-            InstalledAppsOverviewScreen(
+            InstalledAppsScreen(
+                windowSizeClass = windowSizeClass,
                 onNavigate = {
                     navController.navigate(it)
                 },
