+++ features/apps/ui/src/main/kotlin/com/nasdroid/apps/ui/installed/overview/InstalledAppsOverviewScreen.kt
@@ -1,4 +1,4 @@
-package com.nasdroid.apps.ui.installed
+package com.nasdroid.apps.ui.installed.overview
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -8,9 +8,8 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.GetApp
@@ -35,18 +34,18 @@ import androidx.compose.ui.unit.dp
 import com.boswelja.menuprovider.MenuItem
 import com.boswelja.menuprovider.ProvideMenuItems
 import com.nasdroid.apps.ui.R
-import com.nasdroid.apps.ui.installed.item.ApplicationOverviewItem
+import com.nasdroid.apps.ui.installed.overview.item.ApplicationOverviewItem
 import org.koin.androidx.compose.koinViewModel
 
 /**
- * A screen for displaying and allowing management of apps installed on the system.
+ * A screen for displaying apps installed on the system.
  */
 @Composable
 fun InstalledAppsOverviewScreen(
     onNavigate: (route: String) -> Unit,
     modifier: Modifier = Modifier,
     contentPadding: PaddingValues = PaddingValues(),
-    viewModel: InstalledAppsViewModel = koinViewModel()
+    viewModel: InstalledAppsOverviewViewModel = koinViewModel()
 ) {
     ProvideMenuItems(
         MenuItem(
@@ -62,11 +61,9 @@ fun InstalledAppsOverviewScreen(
     var deletingApp by rememberSaveable { mutableStateOf<String?>(null) }
 
     Box(modifier) {
-        LazyVerticalGrid(
-            columns = GridCells.Adaptive(240.dp),
+        LazyColumn(
             contentPadding = contentPadding,
             verticalArrangement = Arrangement.spacedBy(8.dp),
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
         ) {
             installedApps?.let { apps ->
                 items(
